Endpoint para buscar reteios de votos e scheduler para retaear votos …

+++ src/main/java/br/pereira/votous/api/v1/dto/VotoInputDTO.java
@@ -8,7 +8,7 @@ import lombok.Setter;
 @Setter
 public class VotoInputDTO {
 
-    private Integer cdPauta;
+    private Long cdPauta;
     private Integer cdSessao;
     private String cpf;
     private VotoEnum voto;
